requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.13
django-ckeditor==6.7.1
django-js-asset==2.2.0
pillow==10.3.0
pytz==2023.3.post1
sqlparse==0.5.0
typing_extensions==4.11.0

django-recaptcha
